pyproject.toml
[tool.poetry]
name = "payroll"
version = "1.0.0"
description = "A command-line tool to help Phalanx Software compute and execute their payroll"
authors = ["Nicky Buttigieg <nicholas.buttigieg@phalanx.software>"]

[tool.poetry.dependencies]
python = "^3.8"
pyyaml = "^5.3.1"
python-dateutil = "^2.8.1"
pydantic = "^1.6.1"
click = "^7.1.2"
jinja2 = "^2.11.2"
pdfkit = "^0.6.1"
python-dotenv = "^0.14.0"
wkhtmltopdf = "^0.2"
py-moneyed = "^1.1"
pytest = "^6.2.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.poetry.scripts]
payroll = "payroll.main:main"